+++ esibd/examples/__init__.py
@@ -2,4 +2,5 @@
 
 from .calculator.calculator_plugin import Calculator  # noqa: F401
 from .custom_device.custom_device import CustomDevice  # noqa: F401
+from .custom_device.custom_device2 import CustomDevice2  # noqa: F401
 from .custom_plugin.custom_plugin import CustomPlugin  # noqa: F401